+++ src/aux-server/aux-web/aux-player/scene/PlayerGame.ts
@@ -30,6 +30,13 @@ import {
     DEFAULT_INVENTORY_VISIBLE,
     DEFAULT_PORTAL_SHOW_FOCUS_POINT,
     DEFAULT_PORTAL_DISABLE_CANVAS_TRANSPARENCY,
+    BufferSoundAction,
+    hasValue,
+    BotAction,
+    enqueueAsyncResult,
+    enqueueAsyncError,
+    PlaySoundAction,
+    CancelSoundAction,
 } from '@casual-simulation/aux-common';
 import {
     baseAuxAmbientLight,
@@ -44,7 +51,10 @@ import { CameraRigControls } from '../../shared/interaction/CameraRigControls';
 import { AuxBotVisualizer } from '../../shared/scene/AuxBotVisualizer';
 import { ItemDimension } from '../ItemDimension';
 import { DimensionItem } from '../DimensionItem';
-import { getBotsStateFromStoredAux } from '@casual-simulation/aux-vm';
+import {
+    getBotsStateFromStoredAux,
+    Simulation,
+} from '@casual-simulation/aux-vm';
 import { GameAudio } from '../../shared/scene/GameAudio';
 
 export class PlayerGame extends Game {
@@ -425,7 +435,11 @@ export class PlayerGame extends Game {
                 } else if (e.type === 'import_aux') {
                     this.importAUX(sim, e.url);
                 } else if (e.type === 'play_sound') {
-                    this.playAudio(e.url);
+                    this.playAudio(sim, e);
+                } else if (e.type === 'buffer_sound') {
+                    this.bufferAudio(sim, e);
+                } else if (e.type === 'cancel_sound') {
+                    this.cancelAudio(sim, e);
                 } else if (e.type === 'enable_ar') {
                     if (e.enabled) {
                         this.startAR();
@@ -443,10 +457,45 @@ export class PlayerGame extends Game {
         );
     }
 
-    playAudio(url: string) {
-        if (url === null) return;
+    playAudio(sim: Simulation, event: PlaySoundAction) {
+        if (event.url === null) return;
 
-        this.audio.playFromUrl(url);
+        this.audio.playFromUrl(event.url, event.soundID).then(
+            () => {
+                let list = [] as BotAction[];
+                enqueueAsyncResult(list, event, event.soundID, false);
+                sim.helper.transaction(...list);
+            },
+            err => {
+                let list = [] as BotAction[];
+                enqueueAsyncError(list, event, err);
+                sim.helper.transaction(...list);
+            }
+        );
+    }
+
+    bufferAudio(sim: Simulation, event: BufferSoundAction) {
+        if (!hasValue(event.url)) return;
+
+        this.audio.bufferFromUrl(event.url).then(
+            () => {
+                let list = [] as BotAction[];
+                enqueueAsyncResult(list, event, null, false);
+                sim.helper.transaction(...list);
+            },
+            err => {
+                let list = [] as BotAction[];
+                enqueueAsyncError(list, event, err);
+                sim.helper.transaction(...list);
+            }
+        );
+    }
+
+    cancelAudio(sim: BrowserSimulation, event: CancelSoundAction) {
+        this.audio.cancelSound(event.soundID);
+        let list = [] as BotAction[];
+        enqueueAsyncResult(list, event, null, false);
+        sim.helper.transaction(...list);
     }
 
     private simulationRemoved(sim: BrowserSimulation) {