+++ src/aux-server/aux-web/shared/vue-components/MonacoTagEditor/MonacoTagEditor.vue
@@ -1,12 +1,22 @@
 <template>
     <div class="editor-wrapper">
-        <div class="editor-breadcrumbs">
+        <div class="editor-breadcrumbs" :class="{ 'show-resize': showResize }">
             <bot-tag
                 :tag="tag"
                 :isScript="isScript"
                 :isFormula="isFormula"
                 :allowCloning="false"
             ></bot-tag>
+            <div class="editor-toggles">
+                <md-button @click="makeNormalTag()" class="md-dense" :class="{ active: !isScript }">
+                    <md-tooltip>Make Normal Tag</md-tooltip>
+                    <span class="hashtag">#</span>
+                </md-button>
+                <md-button @click="makeScriptTag()" class="md-dense" :class="{ active: isScript }">
+                    <md-tooltip>Make Listen Tag</md-tooltip>
+                    <span class="at-symbol">@</span>
+                </md-button>
+            </div>
             <div class="editor-errors">
                 <div v-if="scriptErrors.length === 0">
                     No Errors